+++ js/const.js
@@ -1,0 +1,30 @@
+export const TYPES_OF_BUILDING = [
+  "palace",
+  "flat",
+  "house",
+  "bungalow",
+  "hotel",
+];
+export const CHECK_TIMES = ["12:00", "13:00", "14:00"];
+export const FEATURES = [
+  "wifi",
+  "dishwasher",
+  "parking",
+  "washer",
+  "elevator",
+  "conditioner",
+];
+export const PHOTOS = [
+  "https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg",
+  "https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg",
+  "https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg",
+];
+export const TITLES = ["Квартира", "Дом", "Лофт", "Дача"];
+export const DESCRIPTIONS = [
+  "Шикарноый вид из окна",
+  "Отличный ремонт",
+  "Есть вся мебель",
+  "Есть вся техника для уютного проживания",
+  "Хорошие соседи",
+];
+export const RANDOM_OFFER_COUNT = 10;
